+++ NewGameEngine/NewGameEngine.vcxproj
@@ -110,11 +110,13 @@
     <ClCompile Include="Source\Graphics\Bindable\VertexShader.cpp" />
     <ClCompile Include="Source\Graphics\ChiliCamera.cpp" />
     <ClCompile Include="Source\Graphics\Drawable\Drawable.cpp" />
-    <ClCompile Include="Source\Graphics\Drawable\Particle.cpp" />
     <ClCompile Include="Source\Graphics\Drawable\SolidSphere.cpp" />
     <ClCompile Include="Source\Graphics\Drawable\WrapperSolidSphere.cpp" />
     <ClCompile Include="Source\Graphics\Graphics.cpp" />
     <ClCompile Include="Source\Graphics\NewCamera.cpp" />
+    <ClCompile Include="Source\Graphics\Particle_Systems\Particle.cpp" />
+    <ClCompile Include="Source\Graphics\Particle_Systems\ParticleSystem.cpp" />
+    <ClCompile Include="Source\Graphics\Particle_Systems\ParticleWrapper.cpp" />
     <ClCompile Include="Source\Graphics\Surface.cpp" />
     <ClCompile Include="Source\GUI\GUImanager.cpp" />
     <ClCompile Include="Source\GUI\GUIwrap.cpp" />
@@ -160,13 +162,15 @@
     <ClInclude Include="Source\Graphics\Drawable\Drawable.h" />
     <ClInclude Include="Source\Graphics\Drawable\IndexedTriangleList.h" />
     <ClInclude Include="Source\Graphics\Drawable\IndexedTriangleList.h" />
-    <ClInclude Include="Source\Graphics\Drawable\Particle.h" />
     <ClInclude Include="Source\Graphics\Drawable\SolidSphere.h" />
     <ClInclude Include="Source\Graphics\Drawable\Sphere.h" />
     <ClInclude Include="Source\Graphics\Drawable\WrapperSolidSphere.h" />
     <ClInclude Include="Source\Graphics\EMaths.h" />
     <ClInclude Include="Source\Graphics\Graphics.h" />
     <ClInclude Include="Source\Graphics\NewCamera.h" />
+    <ClInclude Include="Source\Graphics\Particle_Systems\Particle.h" />
+    <ClInclude Include="Source\Graphics\Particle_Systems\ParticleSystem.h" />
+    <ClInclude Include="Source\Graphics\Particle_Systems\ParticleWrapper.h" />
     <ClInclude Include="Source\Graphics\Surface.h" />
     <ClInclude Include="Source\Graphics\Vertex.h" />
     <ClInclude Include="Source\GUI\GUImanager.h" />
