+++ NewGameEngine/Source/ObjectFactory.cpp
@@ -1,7 +1,7 @@
 #include "ObjectFactory.h"
 #include "SceneManager.h"
 #include "../Source/Graphics/Drawable/WrapperSolidSphere.h"
-#include "../Source/Graphics/Drawable/Particle.h"
+#include "../Source/Graphics/Particle_Systems/Particle.h"
 
 void ObjectFactory::SetUpObjectManager(std::shared_ptr<ObjectManager> om)
 {
@@ -36,22 +36,22 @@ void ObjectFactory::AddSphere(Graphics& gfx, float pos[3], float radius, float l
 void ObjectFactory::AddParticleSystem(Graphics& gfx, DirectX::XMFLOAT3 Velocity, DirectX::XMFLOAT3 VelocityVariation, DirectX::XMFLOAT4 ColorBegin, DirectX::XMFLOAT4 ColorEnd, float SizeBegin, float SizeEnd, float SizeVariation, float LifeTime, const char* vs, const char* ps, float pos[3], const char* gs)
 {
 	
-	if (this->om) {
-		this->om->AddObject(std::make_shared<ParticleSystem>(
-			gfx,
-			Velocity,
-			VelocityVariation,
-			ColorBegin,
-			ColorEnd,
-			SizeBegin,
-			SizeEnd,
-			SizeVariation,
-			LifeTime,
-			vs,
-			ps,
-			pos,
-			gs
-			)
-		);
-	}
+	//if (this->om) {
+	//	this->om->AddObject(std::make_shared<ParticleSystem>(
+	//		gfx,
+	//		Velocity,
+	//		VelocityVariation,
+	//		ColorBegin,
+	//		ColorEnd,
+	//		SizeBegin,
+	//		SizeEnd,
+	//		SizeVariation,
+	//		LifeTime,
+	//		vs,
+	//		ps,
+	//		pos,
+	//		gs
+	//		)
+	//	);
+	//}
 }
