+++ NewGameEngine/Source/ObjectManager.h
@@ -4,14 +4,15 @@
 #include <memory>
 #include "SceneManager.h"
 #include "SceneObject.h"
-#include "Graphics/Drawable/Particle.h"
+#include "Graphics/Particle_Systems/ParticleSystem.h"
+// #include "Graphics/Particle_Systems/Particle.h"
 
 
 class ObjectManager
 {
 public:
 	ObjectManager() {};
-	void AddObject(std::shared_ptr<SceneObject> obj);		// ObjectFactory related method
+	void AddObject(std::shared_ptr<SceneObject> obj);		
 	void SetLevelManager(std::shared_ptr<SceneManager> sm) { if(!this->sm) this->sm = std::move(sm); };
 	void UpdateAll(Graphics& gfx, DirectX::XMMATRIX viewMatrix, float musParams[3], float timeFrame);						// Update all objects in the pObjectList
 	void RenderAll(Graphics& gfx);										// Render all objects in the pObjectList
@@ -25,6 +26,7 @@ private:
 	template<class T> std::shared_ptr<T> QueryObject();
 	template<class T> std::list<std::shared_ptr<T>> QueryObjectList();
 	std::list<std::shared_ptr<SceneObject>> pObjectList;
+	std::list<std::shared_ptr<ParticleSystem>> pPSList;
 	std::shared_ptr<SceneManager> sm;
 };
