+++ NewGameEngine/Source/SceneObject.h
@@ -20,7 +20,13 @@ public:
 	virtual DirectX::XMFLOAT3 getPos() { return this->pos; };
 	bool IsActive() { return this->active; };
 	virtual void Bind(Graphics& gfx, DirectX::FXMMATRIX view, float musParams[3], float timeFrame) const  = 0;
-	
+	void SetPos(DirectX::XMFLOAT3 pos);
+	void SetSize(DirectX::XMFLOAT3 size);
+	void adjustSize(DirectX::XMFLOAT3 dSize);
+	void adjustPosition(DirectX::XMFLOAT3 dPos);
+	DirectX::XMFLOAT3 GetPos();
+	DirectX::XMFLOAT3 GetSize();
+
 private:
 	virtual void _Reset() = 0;
 
@@ -29,6 +35,8 @@ protected:
 	bool active;
 	float initPos[3];
 	DirectX::XMFLOAT3 pos;											// position of the object
+	DirectX::XMFLOAT3 size;											// position of the object
+	
 	std::vector<std::shared_ptr<Drawable>> meshes;
 };
